Make sure visualizations can display with updated libraries.
Separated out the function definitions from the line calling the
function (the point of the function is that it doesn't have to be
repeated, after all). This also allows `%matplotlib widget` to be
used on the function-calling cells, where they're needed to prevent
IPython errors.

Updated the TSNE functions for changed parameter names and required
values.

Placed `import random` up with the other import statements, so it
doesn't need to be called again and again.